+++ app/src/androidTest/java/woowacourse/movie/MovieReservationActivityTest.kt
@@ -18,6 +18,7 @@ import org.junit.Test
 import org.junit.jupiter.api.DisplayName
 import woowacourse.movie.fixtures.fakeContext
 import woowacourse.movie.fixtures.movie
+import woowacourse.movie.view.MovieReservationActivity
 
 class MovieReservationActivityTest {
     private lateinit var intent: Intent
