+++ app/src/androidTest/java/woowacourse/movie/MovieReservationResultActivityTest.kt
@@ -13,6 +13,8 @@ import org.junit.Test
 import org.junit.jupiter.api.DisplayName
 import woowacourse.movie.fixtures.fakeContext
 import woowacourse.movie.fixtures.ticket
+import woowacourse.movie.view.MovieReservationActivity
+import woowacourse.movie.view.MovieReservationResultActivity
 
 class MovieReservationResultActivityTest {
     private lateinit var intent: Intent
