+++ app/src/main/res/layout/activity_movie_reservation.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MovieReservationActivity">
+    tools:context=".view.MovieReservationActivity">
 
     <ImageView
         android:id="@+id/poster"
